+++ src/lambdaisland/makina/test.clj
@@ -11,11 +11,15 @@
 
 (def ^:dynamic *app* nil)
 
-(defn init-sys [opts]
-  (-> (assoc opts :profile :test)
-      app/create*
-      app/load*
-      (app/start* (:keys opts))))
+(defn init-app [opts]
+  (let [app (-> (assoc opts :profile :test)
+                app/create*
+                app/load*
+                (app/start* (:keys opts)))]
+    (when-let [e (app/error* app)]
+      (app/stop* app)
+      (throw e))
+    app))
 
 (defmacro with-app
   "Start a system, bind to `*app*`, evaluate `body`, tear the system down again
@@ -23,7 +27,7 @@
   - `opts` - see [[lambdaisland.makina.application/create]]
   "
   [opts & body]
-  `(binding [*app* (init-sys ~opts)]
+  `(binding [*app* (init-app ~opts)]
      (let [res# (do ~@body)]
        (app/stop* *app*)
        res#)))
@@ -51,41 +55,16 @@
                               ks-or-opts
                               {:keys ks-or-opts}))]
        (fn [t]
-         (with-system opts (t)))))))
+         (with-app opts (t)))))))
 
 (defn component
   "Get a single component value"
   [k]
   (sys/component (:makina/system *app*) k))
 
-(defn wrap-run
-  "Hook for use with Kaocha
-
-  Configure as a \"wrap-run\" hook, it will ensure during the entire test run a
-  app/system is available in `*app*`. Settings (as
-  per [[lambdaisland.makina.app/create]]) can be configured inside `tests.edn`
-  using the `:makina/settings` key. Either set this to a map (e.g.
-  `:makina/settings {:prefix \"my-app\"}`), or use a symbol to point at a var
-  that can provide the settings. The var should resolve to a map or function.
-
-  ```
-  ;; tests.edn
-  #kaocha/v1
-  {:plugins [:hooks]
-   :kaocha.hooks/wrap-run [lambdaisland.makina.test/wrap-run]
-   :makina/settings my-app.config.makina-opts}
-  ```
-  "
-  [run-fn test-plan]
-  (let [settings (:makina/settings test-plan)
-        settings (cond-> settings (symbol? settings) (-> requiring-resolve deref))
-        settings (if (fn? settings) (settings) settings)]
-    (with-app settings
-      (run-fn))))
-
 (defn start!
   "Starts an app/system and binds `*app*` permanently, meant for REPL use"
   ([opts]
-   (alter-var-root #'*app* (constantly (init-sys))))
+   (alter-var-root #'*app* (constantly (init-app opts))))
   ([opts ks]
    (start! (assoc opts :keys ks))))